% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\docType{data}
\name{Browser}
\alias{Browser}
\title{Browser base class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Browser
}
\description{
Base class for browsers like Chrome, Chromium, etc. Defines the interface
used by various browser implementations. It can represent a local browser
process or one running remotely.
}
\details{
The \code{initialize()} method of an implementation should set private$host
and private$port. If the process is local, the \code{initialize()} method
should also set private$process.
}
\keyword{datasets}
