% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{chrome_cache}
\alias{chrome_cache}
\alias{chromote_cache}
\alias{chromote_cache_path}
\alias{chrome_cache_path}
\alias{chrome_cache_path_installed}
\alias{chrome_cache_remove}
\title{Chromote cache path helpers}
\usage{
chromote_cache_path(...)

chrome_cache_path(...)

chrome_cache_path_installed(
  version = "latest",
  binary = "chrome",
  platform = NULL
)

chrome_cache_remove(version, binary, platform = NULL, ask = TRUE)
}
\arguments{
\item{...}{Additional path parts.}

\item{version}{A character string specifying the version to use. Default is
\code{"latest-installed"}, which uses the latest version you have installed via
chromote. To use the current latest version, downloading the binary if
necessary. If you specify a partial version, e.g. \code{"132"}, chromote will
find the most recent release matching that version. Use
\code{version = "system"} to revert back to using the system-installed Chrome.}

\item{binary}{A character string specifying which binary to list. Defaults to
\code{"all"} to show all binaries, or can be one or more of of \code{"chrome"},
\code{"chrome-headless-shell"}, or \code{"chromedriver"}.}

\item{platform}{A character string specifying the platform(s) to list. If
\code{NULL} (default), the platform will be automatically detected, or if
\code{"all"}, then binaries for all platforms will be listed.}

\item{ask}{Whether to ask before removing files.}
}
\value{
A character vector of paths.
}
\description{
These functions help interact with \pkg{chromote}'s cache, primarily used to
cache Chrome for Testing binaries:
\itemize{
\item \code{chromote_cache_path()}: Returns the path to the general \pkg{chromote}
cache.
\item \code{chrome_cache_path()}: Returns the path to the cache used for Chrome
binaries (inside the general cache).
\item \code{chrome_cache_path_installed()}: Returns a path or paths to specific Chrome
binaries in the cache.
\item \code{chromem_cache_remove()}: Remove specific versions and binaries from the
Chrome cache. The \code{version}, \code{binary} and \code{platform} arguments can each
take \code{"all"} to remove all installed copies of that version, binary or
platform.
}
}
