% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{chrome_use_version}
\alias{chrome_use_version}
\title{Use a specific version of Chrome or related binaries}
\usage{
chrome_use_version(
  version = "latest-installed",
  binary = c("chrome", "chrome-headless-shell", "chromedriver"),
  platform = NULL
)
}
\arguments{
\item{version}{A character string specifying the version to use. Default is
\code{"latest-installed"}, which uses the latest version you have installed via
chromote. To use the current latest version, downloading the binary if
necessary. If you specify a partial version, e.g. \code{"132"}, chromote will
find the most recent release matching that version. Use
\code{version = "system"} to revert back to using the system-installed Chrome.}

\item{binary}{A character string specifying which binary to
use. Must be one of \code{"chrome"}, \code{"chrome-headless-shell"}, or
\code{"chromedriver"}. Default is \code{"chrome"}.}

\item{platform}{A character string specifying the platform. If \code{NULL}
(default), the platform will be automatically detected.}
}
\value{
Sets the \code{CHROMOTE_CHROME} environment variable and invisibly returns
the old value of the \code{CHROMOTE_CHROME} environment variable.
}
\description{
This function downloads and sets up a specific version of Chrome, using the
\href{https://googlechromelabs.github.io/chrome-for-testing}{Google Chrome for Testing builds}
for \code{chrome}, \code{chrome-headless-shell} or \code{chromedriver} for use with
chromote.
}
\details{
This function downloads the specified binary, sets up the necessary
environment, and returns the old CHROMOTE_CHROME environment variable
value. It uses the "known-good-versions" which includes all Google Chrome
for Testing versions from
https://googlechromelabs.github.io/chrome-for-testing
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Use the latest version of Chrome
chrome_use_version()

# Use a specific version of chrome-headless-shell
chrome_use_version("114.0.5735.90", binary = "chrome-headless-shell")
\dontshow{\}) # examplesIf}
}
