% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\docType{data}
\name{ChromeRemote}
\alias{ChromeRemote}
\title{Class representing a remote Chrome process}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ChromeRemote
}
\description{
Class representing a remote Chrome process
}
\keyword{datasets}
