% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote-package.R
\name{chromote-options}
\alias{chromote-options}
\title{chromote Options}
\description{
These options and environment variables that are used by chromote. Options
are lowercase and can be set with \code{options()}. Environment variables are
uppercase and can be set in an \code{.Renviron} file, with \code{Sys.setenv()}, or in
the shell or process running R. If both an option or environment variable are
supported, chromote will use the option first.
\itemize{
\item \code{CHROMOTE_CHROME} \cr
Path to the Chrome executable. If not set, chromote will
attempt to find and use the system installation of Chrome.
\item \code{chromote.headless}, \code{CHROMOTE_HEADLESS} \cr
Headless mode for Chrome. Can be \code{"old"} or \code{"new"}. See
\href{https://developer.chrome.com/docs/chromium/new-headless}{Chrome Headless mode}
for more details.
\item \code{chromote.timeout} \cr
Timeout (in seconds) for Chrome to launch or connect. Default is \code{10}.
\item \code{chromote.launch.echo_cmd} \cr
Echo the command used to launch Chrome to the console for debugging.
Default is \code{FALSE}.
}
}
