% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote.R
\name{Chromote}
\alias{Chromote}
\title{Chromote class}
\description{
Chromote class

Chromote class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Chromote$new()}}
\item \href{#method-view}{\code{Chromote$view()}}
\item \href{#method-get_auto_events}{\code{Chromote$get_auto_events()}}
\item \href{#method-get_child_loop}{\code{Chromote$get_child_loop()}}
\item \href{#method-wait_for}{\code{Chromote$wait_for()}}
\item \href{#method-new_session}{\code{Chromote$new_session()}}
\item \href{#method-get_sessions}{\code{Chromote$get_sessions()}}
\item \href{#method-register_session}{\code{Chromote$register_session()}}
\item \href{#method-send_command}{\code{Chromote$send_command()}}
\item \href{#method-invoke_event_callbacks}{\code{Chromote$invoke_event_callbacks()}}
\item \href{#method-debug_messages}{\code{Chromote$debug_messages()}}
\item \href{#method-debug_log}{\code{Chromote$debug_log()}}
\item \href{#method-url}{\code{Chromote$url()}}
\item \href{#method-is_active}{\code{Chromote$is_active()}}
\item \href{#method-get_browser}{\code{Chromote$get_browser()}}
\item \href{#method-stop}{\code{Chromote$stop()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$new(browser = Chrome$new(), multi_session = TRUE, auto_events = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view"></a>}}
\if{latex}{\out{\hypertarget{method-view}{}}}
\subsection{Method \code{view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$view()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_auto_events"></a>}}
\if{latex}{\out{\hypertarget{method-get_auto_events}{}}}
\subsection{Method \code{get_auto_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$get_auto_events()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_child_loop"></a>}}
\if{latex}{\out{\hypertarget{method-get_child_loop}{}}}
\subsection{Method \code{get_child_loop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$get_child_loop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wait_for"></a>}}
\if{latex}{\out{\hypertarget{method-wait_for}{}}}
\subsection{Method \code{wait_for()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$wait_for(p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new_session"></a>}}
\if{latex}{\out{\hypertarget{method-new_session}{}}}
\subsection{Method \code{new_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$new_session(width = 992, height = 774, targetId = NULL, wait_ = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sessions"></a>}}
\if{latex}{\out{\hypertarget{method-get_sessions}{}}}
\subsection{Method \code{get_sessions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$get_sessions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_session"></a>}}
\if{latex}{\out{\hypertarget{method-register_session}{}}}
\subsection{Method \code{register_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$register_session(session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send_command"></a>}}
\if{latex}{\out{\hypertarget{method-send_command}{}}}
\subsection{Method \code{send_command()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$send_command(
  msg,
  callback = NULL,
  error = NULL,
  timeout = NULL,
  sessionId = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invoke_event_callbacks"></a>}}
\if{latex}{\out{\hypertarget{method-invoke_event_callbacks}{}}}
\subsection{Method \code{invoke_event_callbacks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$invoke_event_callbacks(event, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-debug_messages"></a>}}
\if{latex}{\out{\hypertarget{method-debug_messages}{}}}
\subsection{Method \code{debug_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$debug_messages(value = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-debug_log"></a>}}
\if{latex}{\out{\hypertarget{method-debug_log}{}}}
\subsection{Method \code{debug_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$debug_log(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-url"></a>}}
\if{latex}{\out{\hypertarget{method-url}{}}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$url(path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_active"></a>}}
\if{latex}{\out{\hypertarget{method-is_active}{}}}
\subsection{Method \code{is_active()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$is_active()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_browser"></a>}}
\if{latex}{\out{\hypertarget{method-get_browser}{}}}
\subsection{Method \code{get_browser()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$get_browser()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chromote$stop()}\if{html}{\out{</div>}}
}

}
}
