% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{chromote_info}
\alias{chromote_info}
\title{Show information about the chromote package and Chrome browser}
\usage{
chromote_info()
}
\value{
A list containing the following elements:
\describe{
\item{os}{The operating system platform.}
\item{version_r}{The version of R.}
\item{version_chromote}{The version of the chromote package.}
\item{envvar}{The value of the \code{CHROMOTE_CHROME} environment variable.}
\item{path}{The path to the Chrome browser.}
\item{args}{A vector of Chrome arguments.}
\item{version}{The version of Chrome (if verification is successful).}
\item{error}{The error message (if verification fails).}
\item{.check}{A list with the status and output of the Chrome verification.}
}
}
\description{
This function gathers information about the operating system, R version,
chromote package version, environment variables, Chrome path, and Chrome
arguments. It also verifies the Chrome installation and retrieves its version.
}
\examples{
chromote_info()

}
