% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote_session.R
\docType{data}
\name{ChromoteSession}
\alias{ChromoteSession}
\title{ChromoteSession class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ChromoteSession
}
\description{
ChromoteSession class
}
\keyword{datasets}
