% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{default_browser}
\alias{default_browser}
\title{Returns a default browser object}
\usage{
default_browser()
}
\description{
This function first calls \code{getOption("chromote.browser")} and sees
if the value is a function; if so, it invokes that function and returns the
value. If the value not a function (or is not unset), this function calls
\code{\link{chrome}()} and returns the value.
}
