% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%...>\%}
\alias{\%...!\%}
\alias{\%...T>\%}
\alias{\%...T!\%}
\alias{\%>\%}
\alias{\%T>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%T>\%}}}

  \item{promises}{\code{\link[promises]{\%...>\%}}, \code{\link[promises]{\%...!\%}}, \code{\link[promises]{\%...T>\%}}, \code{\link[promises]{\%...T!\%}}}
}}

