% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\docType{data}
\name{Chrome}
\alias{Chrome}
\title{Class representing a local Chrome process}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Chrome
}
\description{
Class representing a local Chrome process
}
\keyword{datasets}
