% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromium.R
\name{chrome}
\alias{chrome}
\title{Return a Chrome object}
\usage{
chrome()
}
\description{
This will start a Chrome process if necessary. If one is already running,
the object representing that process will be returned.
}
