% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote-package.R
\docType{package}
\name{chromote-package}
\alias{chromote}
\alias{chromote-package}
\title{chromote: Headless Chrome Web Browser Interface}
\description{
An implementation of the 'Chrome DevTools Protocol', for controlling a headless Chrome web browser.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/chromote/}
  \item \url{https://github.com/rstudio/chromote}
  \item Report bugs at \url{https://github.com/rstudio/chromote/issues}
}

}
\author{
\strong{Maintainer}: Winston Chang \email{winston@posit.co}

Authors:
\itemize{
  \item Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Garrick Aden-Buie \email{garrick@posit.co} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
