% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote_session.R
\name{ChromoteSession}
\alias{ChromoteSession}
\title{ChromoteSession class}
\description{
ChromoteSession class

ChromoteSession class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ChromoteSession$new()}}
\item \href{#method-init_promise}{\code{ChromoteSession$init_promise()}}
\item \href{#method-close}{\code{ChromoteSession$close()}}
\item \href{#method-view}{\code{ChromoteSession$view()}}
\item \href{#method-is_active}{\code{ChromoteSession$is_active()}}
\item \href{#method-new_session}{\code{ChromoteSession$new_session()}}
\item \href{#method-send_command}{\code{ChromoteSession$send_command()}}
\item \href{#method-get_parent}{\code{ChromoteSession$get_parent()}}
\item \href{#method-get_session_id}{\code{ChromoteSession$get_session_id()}}
\item \href{#method-get_child_loop}{\code{ChromoteSession$get_child_loop()}}
\item \href{#method-wait_for}{\code{ChromoteSession$wait_for()}}
\item \href{#method-get_auto_events}{\code{ChromoteSession$get_auto_events()}}
\item \href{#method-debug_log}{\code{ChromoteSession$debug_log()}}
\item \href{#method-invoke_event_callbacks}{\code{ChromoteSession$invoke_event_callbacks()}}
\item \href{#method-mark_closed}{\code{ChromoteSession$mark_closed()}}
\item \href{#method-screenshot}{\code{ChromoteSession$screenshot()}}
\item \href{#method-screenshot_pdf}{\code{ChromoteSession$screenshot_pdf()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ChromoteSession} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$new(
  parent = default_chromote_object(),
  width = 992,
  height = 744,
  targetId = NULL,
  wait_ = TRUE,
  auto_events = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{\code{\link{Chromote}} object to use; defaults to
\code{\link[=default_chromote_object]{default_chromote_object()}}}

\item{\code{width}}{Width, in pixels, of the \code{Target} to create if \code{targetId} is
\code{NULL}}

\item{\code{height}}{Height, in pixels, of the \code{Target} to create if \code{targetId} is
\code{NULL}}

\item{\code{targetId}}{\href{https://chromedevtools.github.io/devtools-protocol/tot/Target}{Target}
ID of an existing target to attach to. When a \code{targetId} is provided, the
\code{width} and \code{height} arguments are ignored. If NULL (the default) a new
target is created and attached to, and the \code{width} and \code{height}
arguments determine its viewport size.}

\item{\code{wait_}}{If \code{FALSE}, return a \code{\link[promises:promise]{promises::promise()}} of a new
\code{ChromoteSession} object. Otherwise, block during initialization, and
return a \code{ChromoteSession} object directly.}

\item{\code{auto_events}}{If \code{NULL} (the default), use the \code{auto_events} setting
from the parent \code{Chromote} object. If \code{TRUE}, enable automatic
event enabling/disabling; if \code{FALSE}, disable automatic event
enabling/disabling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ChromoteSession} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_promise"></a>}}
\if{latex}{\out{\hypertarget{method-init_promise}{}}}
\subsection{Method \code{init_promise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$init_promise()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$close(wait_ = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view"></a>}}
\if{latex}{\out{\hypertarget{method-view}{}}}
\subsection{Method \code{view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$view()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_active"></a>}}
\if{latex}{\out{\hypertarget{method-is_active}{}}}
\subsection{Method \code{is_active()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$is_active()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new_session"></a>}}
\if{latex}{\out{\hypertarget{method-new_session}{}}}
\subsection{Method \code{new_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$new_session(
  width = 992,
  height = 774,
  targetId = NULL,
  wait_ = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send_command"></a>}}
\if{latex}{\out{\hypertarget{method-send_command}{}}}
\subsection{Method \code{send_command()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$send_command(
  msg,
  callback = NULL,
  error = NULL,
  timeout = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parent"></a>}}
\if{latex}{\out{\hypertarget{method-get_parent}{}}}
\subsection{Method \code{get_parent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$get_parent()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_session_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_session_id}{}}}
\subsection{Method \code{get_session_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$get_session_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_child_loop"></a>}}
\if{latex}{\out{\hypertarget{method-get_child_loop}{}}}
\subsection{Method \code{get_child_loop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$get_child_loop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wait_for"></a>}}
\if{latex}{\out{\hypertarget{method-wait_for}{}}}
\subsection{Method \code{wait_for()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$wait_for(p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_auto_events"></a>}}
\if{latex}{\out{\hypertarget{method-get_auto_events}{}}}
\subsection{Method \code{get_auto_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$get_auto_events()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-debug_log"></a>}}
\if{latex}{\out{\hypertarget{method-debug_log}{}}}
\subsection{Method \code{debug_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$debug_log(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invoke_event_callbacks"></a>}}
\if{latex}{\out{\hypertarget{method-invoke_event_callbacks}{}}}
\subsection{Method \code{invoke_event_callbacks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$invoke_event_callbacks(event, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mark_closed"></a>}}
\if{latex}{\out{\hypertarget{method-mark_closed}{}}}
\subsection{Method \code{mark_closed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$mark_closed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-screenshot"></a>}}
\if{latex}{\out{\hypertarget{method-screenshot}{}}}
\subsection{Method \code{screenshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$screenshot(
  filename = "screenshot.png",
  selector = "html",
  cliprect = NULL,
  region = c("content", "padding", "border", "margin"),
  expand = NULL,
  scale = 1,
  show = FALSE,
  wait_ = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-screenshot_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-screenshot_pdf}{}}}
\subsection{Method \code{screenshot_pdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromoteSession$screenshot_pdf(
  filename = "screenshot.pdf",
  pagesize = "letter",
  margins = 0.5,
  units = c("in", "cm"),
  landscape = FALSE,
  display_header_footer = FALSE,
  print_background = FALSE,
  scale = 1,
  wait_ = TRUE
)}\if{html}{\out{</div>}}
}

}
}
