% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%...>\%}
\alias{\%...!\%}
\alias{\%...T>\%}
\alias{\%...T!\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{promise}
\alias{then}
\alias{catch}
\alias{finally}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:tee]{\%T>\%}}}

  \item{promises}{\code{\link[promises:pipes]{\%...!\%}}, \code{\link[promises:pipes]{\%...>\%}}, \code{\link[promises:pipes]{\%...T!\%}}, \code{\link[promises:pipes]{\%...T>\%}}, \code{\link[promises:then]{catch}}, \code{\link[promises:then]{finally}}, \code{\link[promises]{promise}}, \code{\link[promises]{then}}}
}}

